fix: fix error
- 현재 임시삭제해도 리스트에 남아있는 현상 발생 @kjy-torder
