+++ lib/presentation/const/strings.dart
@@ -7,7 +7,8 @@ class Strings {
   static const DELETE = "DELETE";
   static const CANCEL = "CANCEL";
   static const SAVE = "저장";
-  static const DELETE_CONFIRM_TEXT = "정말 삭제하시겠습니까?";
+  static const TEMP_DELETE_CONFIRM_TEXT = "정말 삭제하시겠습니까?";
+  static const DELETE_CONFIRM_TEXT = "정말 영구 삭제하시겠습니까?";
   static const TEMP_DELETE_WARNING = "*스와이프하면 영구삭제됩니다.";
   static const TEMP_DELETE_TITLE = "휴지통";
   static const DATE_COUNT_UNIT = "개";
