+++ lib/presentation/screen/component/schedule_list_view.dart
@@ -1,12 +1,9 @@
 import 'package:calendar_scheduler/domain/usecase/get_all_schedule.dart';
-import 'package:calendar_scheduler/presentation/const/strings.dart';
-import 'package:calendar_scheduler/presentation/screen/component/schedule.builder.dart';
+import 'package:calendar_scheduler/presentation/screen/component/schedule_item_builder.dart';
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
-
-import '../../../di/locator.dart';
-import '../../../domain/entity/schedule.dart';
-import '../../../domain/usecase/delete_shcedule.dart';
+import '../../const/strings.dart';
+import 'default_component.dart';
 
 class ScheduleListView extends StatelessWidget {
   final DateTime date;
@@ -15,17 +12,43 @@ class ScheduleListView extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final theme = Theme.of(context);
     return Expanded(
       child: Padding(
         padding: const EdgeInsets.symmetric(horizontal: 8.0),
         child: StreamBuilder(
           stream: GetIt.I<GetAllScheduleUsecase>().invoke(date),
           builder: (context, snapshot) {
-            return buildScheduleList(snapshot, context);
+            if (!snapshot.hasData) {
+              return const Center(
+                child: CircularProgressIndicator(),
+              );
+            }
+
+            if (snapshot.hasData && snapshot.data!.isEmpty) {
+              return Center(
+                child: Text(
+                  Strings.EMPTY_SCHEDULE,
+                  style: Theme.of(context).textTheme.titleLarge,
+                ),
+              );
+            }
+            final list = snapshot.requireData;
+
+            return ListView.separated(
+              itemCount: snapshot.data?.length ?? 0,
+              separatorBuilder: (context, index) {
+                return DefaultComponent.defaultSizedBoxWithHeight;
+              },
+              itemBuilder: (context, index) =>
+                  buildScheduleListItem(context, list[index]),
+            );
           },
         ),
       ),
     );
   }
+
+  @override
+  void dispose() {}
 }
+
